pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "1.0.0"
description = ""
authors = [
	"Paul Nikolaevskii <paulnikolaevsky@gmail.com>"
]
packages = [
	{ include = "page_analyzer" },
]

repository = "https://github.com/Utrian/python-project-83"

[tool.poetry.dependencies]
python = "^3.8.1"
gunicorn = "^20.1.0"
Flask = "^2.2.2"
python-dotenv = "^0.21.1"
coverage = "^7.1.0"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
flake8 = "^5.0.4"
psycopg2-binary = "^2.9.5"
validators = "^0.20.0"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.2"
pg-temp = "^0.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
